# api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Ptium API
  version: 0.1.0
  summary: AI presentation generation, administration, and MCP API
  description: |
    Ptium exposes owner-scoped presentation generation, profile and API-key
    management, administrator operations, and a stateless MCP endpoint.

    JSON REST responses use an envelope containing either `data` or `error`
    plus `requestId`. List responses additionally contain `meta`. The `/mcp`
    endpoint uses raw JSON-RPC 2.0 messages and does not use the REST envelope.

    JSON request bodies are limited to 2 MiB, reject unknown top-level fields,
    and must contain exactly one complete JSON value. UUID path parameters are
    validated before their operation handler runs.

    OIDC access tokens and Ptium API keys are both accepted as Bearer
    credentials. Ptium API keys may alternatively be sent in `X-API-Key`.
    Development authentication is disabled by default and uses
    `X-Ptium-Dev-Secret` unless the server is configured with another header.

    The `x-ptium-api-key-scopes` extension on an operation names the scope a
    Ptium API key must grant. Interactive OIDC and development sessions are
    not restricted by API-key scopes. MCP API-key calls always require
    `mcp:use`; each presentation tool additionally requires its documented
    `presentations:read` or `presentations:write` scope.
servers:
  - url: /
    description: Same origin as the Ptium web application
...
security:
  - BearerAuth: []
  - APIKeyHeader: []
  - DevSecret: []
paths:
...
  /api/v1/auth/config:
    parameters:
      - $ref: '#/components/parameters/XRequestID'
    get:
      operationId: getAuthenticationConfig
      summary: Get public authentication configuration
      description: |
        Returns browser-safe OIDC discovery values and the development-auth
        status. Secrets are never returned. OIDC login uses Authorization Code
        with PKCE when `oidcEnabled` is true.
      tags: [Authentication]
      security: []
      responses:
        '200':
          description: Public authentication configuration.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthConfigEnvelope'
  /api/v1/auth/login:
    parameters:
      - $ref: '#/components/parameters/XRequestID'
    post:
      operationId: passwordLogin
      summary: Sign in a local account with a password
      description: |
        Exchanges a username and password for a browser session token. Available
        when `passwordLoginEnabled` is true in the public auth config, which is
        the case once a local account exists — normally the bootstrap
        administrator seeded from the environment.

        Repeated failures back off per client address; a throttled caller
        receives 429 with a `Retry-After` header. The response never reveals
        whether the username exists.
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              additionalProperties: false
              properties:
                username:
                  type: string
                  minLength: 1
                  description: The account's username or email address.
                password:
                  type: string
                  minLength: 1
      responses:
        '200':
          description: A session token and the signed-in account.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: The username or password is incorrect.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: The account is disabled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: This deployment does not accept password sign-in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          description: Too many sign-in attempts from this client.
          headers:
            Retry-After:
              description: Seconds to wait before trying again.
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/password:
    parameters:
      - $ref: '#/components/parameters/XRequestID'
    post:
      operationId: changePassword
      summary: Change the signed-in account's password
      description: |
        Replaces the password after confirming the current one. Every session
        token issued before the change stops working, so the response carries a
        fresh token for the caller while other browsers are signed out.
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              additionalProperties: false
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
                  minLength: 12
                  maxLength: 256
      responses:
        '200':
          description: Password changed; a replacement session token is returned.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionEnvelope'
        '204':
          description: Password changed, but no replacement token could be issued.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: The current password is incorrect.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: The account signs in through the identity provider and has no password.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/token:
    parameters:
      - $ref: '#/components/parameters/XRequestID'
    post:
      operationId: exchangeOIDCToken
      summary: Exchange an OIDC authorization code
      description: |
        Completes the Authorization Code + PKCE exchange on the browser's behalf.
        A confidential client's secret must never reach a single-page app, so
        Ptium presents it to the identity provider itself. The public auth config
        advertises this path as `tokenExchangeUrl` when it is available; a public
        client exchanges the code with the provider directly instead.

        The refresh token the provider returns is deliberately not forwarded.
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [code, redirect_uri, code_verifier]
              properties:
                grant_type:
                  type: string
                  enum: [authorization_code]
                code:
                  type: string
                redirect_uri:
                  type: string
                code_verifier:
                  type: string
      responses:
        '200':
          description: The provider's token response, without any refresh token.
          content:
            application/json:
              schema:
                type: object
                required: [access_token]
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                  expires_in:
                    type: integer
                  id_token:
                    type: string
                  scope:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: The identity provider rejected the exchange.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: This deployment does not exchange tokens server-side.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/logout:
    parameters:
      - $ref: '#/components/parameters/XRequestID'
    post:
      operationId: signOut
      summary: End the browser session
      description: |
        Expires the session cookie. The cookie is HttpOnly, so a browser cannot
        clear it itself. Always succeeds, whether or not the session was still
        valid — signing out must never fail.

        A bearer token is not revoked by this call; discard it client-side.
      tags: [Authentication]
      security: []
      responses:
        '204':
          description: The session cookie has been expired.
  /api/v1/auth/session:
    parameters:
      - $ref: '#/components/parameters/XRequestID'
    post:
      operationId: startSession
      summary: Trade the current identity for a session cookie
      description: |
        Issues a Ptium session for an already authenticated caller and returns it
        as a `Set-Cookie` as well as in the body.

        This is how a workspace stops depending on the identity provider's access
        token, which lives minutes and whose refresh token deliberately never
        reaches the browser. A Ptium session is renewed while its holder keeps
        working, so an active person is not signed out mid-task.

        An API key is refused: a server-to-server credential does not become a
        browser session.
      tags: [Authentication]
      responses:
        '200':
          description: A session token and the signed-in account.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: An API key cannot open a browser session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: This deployment does not issue session cookies.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/me:
    parameters:
      - $ref: '#/components/parameters/XRequestID'
    get:
      operationId: getCurrentUser
      summary: Get the current user and authentication context
      tags: [Authentication]
      responses:
        '200':
          description: Current Ptium identity.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/profile:
    parameters:
      - $ref: '#/components/parameters/XRequestID'
    get:
      operationId: getProfile
      summary: Get the current user's personalization profile
      tags: [Profile]
      x-ptium-api-key-scopes: [profile:read]
      responses:
        '200':
          description: Personalization profile.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      operationId: replaceProfile
      summary: Replace the current user's personalization profile
      description: Omitted fields are stored as empty strings and preferences default to an empty object.
      tags: [Profile]
      x-ptium-api-key-scopes: [profile:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileWrite'
      responses:
        '200':
          description: Updated personalization profile.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      operationId: updateProfile
      summary: Update the current user's personalization profile
      description: |
        This endpoint currently has the same replacement behavior as PUT;
        omitted scalar fields are cleared and omitted preferences become `{}`.
      tags: [Profile]
      x-ptium-api-key-scopes: [profile:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileWrite'
      responses:
        '200':
          description: Updated personalization profile.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/settings:
    parameters:
      - $ref: '#/components/parameters/XRequestID'
    get:
      operationId: getPublicSettings
      summary: Get public branding and generation settings
      description: >-
        Returns only non-sensitive `branding.*` and `generation.*` settings, plus
        `service.version` — the build the server is running. Authentication is not
        required.
      tags: [Settings]
      security: []
      responses:
        '200':
          description: Public runtime settings keyed by dotted setting name.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicSettingsEnvelope'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/assets/{id}:
    parameters:
      - $ref: '#/components/parameters/PresentationID'
      - $ref: '#/components/parameters/XRequestID'
    get:
      operationId: getAsset
      summary: Read an image
      tags: [Presentations]
      responses:
        '200':
          description: The image's bytes.
          headers:
            ETag:
              description: The image's content hash; the bytes never change for an id.
              schema:
                type: string
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteAsset
      summary: Remove an image
      description: |
        A deck that still references the image keeps its slide; the picture is
        simply not drawn, and compiling its source reports the missing name.
      tags: [Presentations]
      responses:
        '204':
          description: The image has been removed.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/presentations/{id}/duplicate:
    parameters:
      - $ref: '#/components/parameters/XRequestID'
      - $ref: '#/components/parameters/PresentationID'
    post:
      operationId: duplicatePresentation
      summary: Duplicate a presentation
      description: Creates an independent copy with new presentation and slide identifiers.
      tags: [Presentations]
      x-ptium-api-key-scopes: [presentations:write]
      responses:
        '201':
          description: Independent copy created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresentationEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: OIDC JWT or ptium_* API key
      description: |
        Send `Authorization: Bearer <credential>`. The credential may be an
        OIDC access-token JWT validated through issuer discovery and rotating
        JWKS, or an opaque Ptium API key beginning with `ptium_`.
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Alternative header for a Ptium `ptium_*` API key. Do not send it
        together with Authorization. The key must grant the scope named by the
        operation's `x-ptium-api-key-scopes` extension.
    DevSecret:
      type: apiKey
      in: header
      name: X-Ptium-Dev-Secret
      description: |
        Development-only shared secret. It is disabled by default, normally
        restricted to loopback clients, and may be renamed by server
        configuration. A Bearer value beginning `dev:` may accompany it to
        select an allowed development identity.
...